添加麻将房间

+++ models/room.go
@@ -4,24 +4,29 @@ import (
 	"sync"
 	"time"
 
+	"github.com/feel-easy/hole-server/consts"
 	"github.com/feel-easy/hole-server/utils/logs"
 )
 
 type Room struct {
 	sync.Mutex
 
-	ID         int           `json:"id"`
-	Type       int           `json:"type"`
-	Game       Game          `json:"roomGame"`
-	State      int           `json:"state"`
-	Users      map[int]*User `json:"users"`
-	Robots     int           `json:"robots"`
-	Creator    int           `json:"creator"`
-	ActiveTime time.Time     `json:"activeTime"`
-	MaxUsers   int           `json:"maxUsers"`
-	Password   string        `json:"password"`
-	EnableChat bool          `json:"enableChat"`
-	Banker     int64         `json:"banker"`
+	ID         int              `json:"id"`
+	Type       consts.GameType  `json:"type"`
+	Game       Game             `json:"roomGame"`
+	State      consts.RoomState `json:"state"`
+	Users      map[int]*User    `json:"users"`
+	Robots     int              `json:"robots"`
+	Creator    int              `json:"creator"`
+	ActiveTime time.Time        `json:"activeTime"`
+	MaxUsers   int              `json:"maxUsers"`
+	Password   string           `json:"password"`
+	EnableChat bool             `json:"enableChat"`
+	Banker     int              `json:"banker"`
+}
+
+func (room *Room) UserNumber() int {
+	return len(room.Users)
 }
 
 func (room *Room) removeUser(user *User) {
@@ -79,8 +84,9 @@ func (room *Room) broadcast(msg string, exclude ...int) {
 
 func (room *Room) delete() {
 	if room != nil {
-		room.Game.delete()
+		if room.Game != nil {
+			room.Game.delete()
+		}
 		delRoom(room.ID)
-		room.Game.delete()
 	}
 }
